Fixed collision reading bug. Made debugging easier.
Turns out that the old collision reading system had not been replaced by
the new one, making unity implementation impossible. Oops. But now it's
fixed. Also added some debugging features to determine what went wrong
if the field didn't load correctly.

+++ FieldExporter/FieldExporter/Components/ComponentPropertiesForm.cs
@@ -98,9 +98,9 @@ namespace FieldExporter.Controls
         /// Returns the value of the mass numeric up down.
         /// </summary>
         /// <returns></returns>
-        public double GetMass()
+        public float GetMass()
         {
-            return Decimal.ToDouble(massNumericUpDown.Value);
+            return (float)Decimal.ToDouble(massNumericUpDown.Value);
         }
 
         /// <summary>
